src/test/java/seedu/resireg/logic/commands/HistoryCommandTest.java
package seedu.resireg.logic.commands;

import static seedu.resireg.logic.commands.CommandTestUtil.assertCommandSuccess;

import org.junit.jupiter.api.Test;

import seedu.resireg.logic.CommandHistory;
import seedu.resireg.model.Model;
import seedu.resireg.model.ModelManager;

public class HistoryCommandTest {
    private CommandHistory history = new CommandHistory();

    private Model model = new ModelManager();
    private Model expectedModel = new ModelManager();

    @Test
    public void execute() {
        assertCommandSuccess(new HistoryCommand(), model, history, HistoryCommand.MESSAGE_NO_HISTORY, expectedModel);

        String firstCommand = "rooms";
        history.add(firstCommand);
        assertCommandSuccess(new HistoryCommand(), model, history,
                String.format(HistoryCommand.MESSAGE_SUCCESS, firstCommand), expectedModel);

        String secondCommand = "invalid";
        String thirdCommand = "clear";
        history.add(secondCommand);
        history.add(thirdCommand);

        String expectedMessage = String.format(HistoryCommand.MESSAGE_SUCCESS,
                String.join("\n", thirdCommand, secondCommand, firstCommand));
        assertCommandSuccess(new HistoryCommand(), model, history, expectedMessage, expectedModel);
    }

}
